modules/opensmile/requirements.txt
numpy<2
hcai-nova-utils
opensmile
